// src/interfaces/candidata.ts
export class Candidata {
  nome: string;
  numero: string;
  cargo: string;
  estado: string;
  foto_url: string;
  descricao: string;
  website: string;
  facebook: string;
  twitter: string;
  instagram: string;

  constructor() {
    this.nome = "";
    this.numero = "";
    this.cargo = "";
    this.estado = "";
    this.foto_url = "";
    this.descricao = "";
    this.website = "";
    this.facebook = "";
    this.twitter = "";
    this.instagram = "";
  }
}